Sheet #1 (Data type - Conditions)

+++ Sheet #1 (Data type - Conditions)/D_Difference.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
    long long int x,y,z,r;
     cin>>x>>y>>z>>r;
-    int difference=(x*y)-(z*r);
+   long long int difference=(x*y)-(z*r);
     cout<<"Difference"<<" = "<<difference;
     // Difference = -10
 }
